+++ src/lbm/D3Q27.hpp
@@ -9,17 +9,15 @@
  * @class D3Q27
  * @brief Class template representing the D3Q27 lattice model.
  *
- * This class template contains the data specific to the D3Q27 lattice model.
- * Accessors are implemented in the CRTP base class LatticeModelBase.
+ * This class template contains the data specific to the `D3Q27` lattice model.
+ * It provides `static constexpr` methods to access the data.
+ * Use the concept `LatticeModel` to restrict template parameters to valid lattice models.
  *
- * @tparam Real The floating-point type.
+ * @tparam T The floating-point type.
  */
-template <std::floating_point Real>
-class D3Q27 : public LatticeModelBase<D3Q27<Real>>
+template <std::floating_point T>
+class D3Q27 : public LatticeModelBase
 {
-public:
-    friend class LatticeModelBase<D3Q27<Real>>;
-
 private:
     static constexpr int dimension_{3};
 
@@ -33,33 +31,54 @@ private:
          {{-1, 1, -1}}, {{1, -1, -1}}, {{-1, -1, -1}}}
     };
 
-    static constexpr std::array<Real, size_> weights_{
-        static_cast<Real>(8) / static_cast<Real>(27),
-        static_cast<Real>(2) / static_cast<Real>(27),
-        static_cast<Real>(2) / static_cast<Real>(27),
-        static_cast<Real>(2) / static_cast<Real>(27),
-        static_cast<Real>(2) / static_cast<Real>(27),
-        static_cast<Real>(2) / static_cast<Real>(27),
-        static_cast<Real>(2) / static_cast<Real>(27),
-        static_cast<Real>(1) / static_cast<Real>(54),
-        static_cast<Real>(1) / static_cast<Real>(54),
-        static_cast<Real>(1) / static_cast<Real>(54),
-        static_cast<Real>(1) / static_cast<Real>(54),
-        static_cast<Real>(1) / static_cast<Real>(54),
-        static_cast<Real>(1) / static_cast<Real>(54),
-        static_cast<Real>(1) / static_cast<Real>(54),
-        static_cast<Real>(1) / static_cast<Real>(54),
-        static_cast<Real>(1) / static_cast<Real>(54),
-        static_cast<Real>(1) / static_cast<Real>(54),
-        static_cast<Real>(1) / static_cast<Real>(54),
-        static_cast<Real>(1) / static_cast<Real>(54),
-        static_cast<Real>(1) / static_cast<Real>(216),
-        static_cast<Real>(1) / static_cast<Real>(216),
-        static_cast<Real>(1) / static_cast<Real>(216),
-        static_cast<Real>(1) / static_cast<Real>(216),
-        static_cast<Real>(1) / static_cast<Real>(216),
-        static_cast<Real>(1) / static_cast<Real>(216),
-        static_cast<Real>(1) / static_cast<Real>(216),
-        static_cast<Real>(1) / static_cast<Real>(216)
+    static constexpr std::array<T, size_> weights_{
+        static_cast<T>(8) / static_cast<T>(27),  static_cast<T>(2) / static_cast<T>(27),
+        static_cast<T>(2) / static_cast<T>(27),  static_cast<T>(2) / static_cast<T>(27),
+        static_cast<T>(2) / static_cast<T>(27),  static_cast<T>(2) / static_cast<T>(27),
+        static_cast<T>(2) / static_cast<T>(27),  static_cast<T>(1) / static_cast<T>(54),
+        static_cast<T>(1) / static_cast<T>(54),  static_cast<T>(1) / static_cast<T>(54),
+        static_cast<T>(1) / static_cast<T>(54),  static_cast<T>(1) / static_cast<T>(54),
+        static_cast<T>(1) / static_cast<T>(54),  static_cast<T>(1) / static_cast<T>(54),
+        static_cast<T>(1) / static_cast<T>(54),  static_cast<T>(1) / static_cast<T>(54),
+        static_cast<T>(1) / static_cast<T>(54),  static_cast<T>(1) / static_cast<T>(54),
+        static_cast<T>(1) / static_cast<T>(54),  static_cast<T>(1) / static_cast<T>(216),
+        static_cast<T>(1) / static_cast<T>(216), static_cast<T>(1) / static_cast<T>(216),
+        static_cast<T>(1) / static_cast<T>(216), static_cast<T>(1) / static_cast<T>(216),
+        static_cast<T>(1) / static_cast<T>(216), static_cast<T>(1) / static_cast<T>(216),
+        static_cast<T>(1) / static_cast<T>(216)
     };
+
+public:
+    using Real = T;
+
+    /**
+     * @brief Returns the spatial dimension D3Q27 (i.e. 3).
+     *
+     * @return The spatial dimension.
+     */
+    static constexpr auto dimension() -> std::size_t;
+
+    /**
+     * @brief Returns the number of discrete velocity vectors for D3Q27 (i.e. 27).
+     *
+     * @return The number of discrete velocity vectors.
+     */
+    static constexpr auto size() -> std::size_t;
+
+    /**
+     * @brief Returns the discrete velocity vectors of D3Q27.
+     *
+     * @return The discrete velocity vectors in lattice units.
+     */
+    static constexpr auto
+    velocities() -> std::array<std::array<int, D3Q27<T>::dimension_>, D3Q27<T>::size_>;
+
+    /**
+     * @brief Returns the quadrature weight for each discrete velocity vector of D3Q27.
+     *
+     * @return The weight for each discrete velocity vector.
+     */
+    static constexpr auto weights() -> std::array<T, D3Q27<T>::size_>;
 };
+
+#include "D3Q27.tpp"