+++ CHANGELOG.md
@@ -8,7 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
-- `LatticeModelBase`: Interface class template for lattice models.
+- `LatticeModelBase`: Base class for lattice models.
+- `LatticeModel`: Concept for constraining a template parameter to valid lattice models.
 - `D1Q3`: Class template for the D1Q3 lattice model.
 - `D2Q9`: Class template for the D2Q9 lattice model.
 - `D3Q27`: Class template for the D3Q27 lattice model.